refactor: add useMarkNotificationsAsSeen hook
